Refactor: ProblemService 를 Interface로 바꾸고 ProblemServiceImpl추가
